+++ custom_components/kress_fleet/const.py
@@ -13,7 +13,7 @@ from homeassistant.const import Platform
 
 DOMAIN: Final = "kress_fleet"
 NAME: Final = "Kress Fleet"
-VERSION: Final = "0.3.2"
+VERSION: Final = "0.3.3"
 
 PLATFORMS: Final = [
     Platform.LAWN_MOWER,